windows: re-align event filter with macos version
