+++ Software/src/SystemSession.cpp
@@ -7,6 +7,7 @@
 
 #include "SystemSession.hpp"
 #if defined(Q_OS_WIN)
+#include "WindowsSession.hpp"
 #elif defined(Q_OS_MACOS)
 #include "MacOSSession.h"
 #endif
@@ -16,6 +17,7 @@ namespace SystemSession {
     EventFilter* EventFilter::create()
     {
 #if defined(Q_OS_WIN)
+		return new WindowsEventFilter();
 #elif defined(Q_OS_MACOS)
         return new MacOSEventFilter();
 #endif
